Added check in load_file_button_click method Check if and update only when new filename was selected

+++ src/controller.py
@@ -16,10 +16,11 @@ class TcController():
         '''Load file button click event handler.
         Call model load_file method and update view'''
         self.model.load_file()
-        self.view.load_file_entry.delete(0, "end")
-        self.view.load_file_entry.insert(0, self.model.load_file_filename)
-        self.view.load_file_status.config(text=f"File loaded, shape: {self.model.df.shape}")
-        self.view.lookup_column_combobox.config(values=self.model.df_columns)
+        if self.model.load_file_filename != self.view.load_file_entry.get():
+            self.view.load_file_entry.delete(0, "end")
+            self.view.load_file_entry.insert(0, self.model.load_file_filename)
+            self.view.load_file_status.config(text=f"File loaded, shape: {self.model.df.shape}")
+            self.view.lookup_column_combobox.config(values=self.model.df_columns)
 
     def save_file_button_click(self):
         '''Save file button click event handler.
